+++ spec/gem_spec.rb
@@ -20,6 +20,7 @@ RSpec.describe RSGem::Gem do
       let(:rakefile) { File.read("./#{gem_name}/Rakefile") }
       let(:travis) { File.read("./#{gem_name}/.travis.yml") }
       let(:expected_travis) { File.read('./lib/rsgem/support/travis.yml') }
+      let(:spec_helper) { File.read("./#{gem_name}/spec/spec_helper.rb") }
 
       it 'creates a new folder' do
         expect(list_files).to include gem_name
@@ -94,6 +95,10 @@ RSpec.describe RSGem::Gem do
 
       it 'does not create a github actions configuration file' do
         expect(File.exist?("./#{gem_name}/.github/workflows")).to eq false
+      end
+
+      it 'adds simplecov configuration in spec helper file' do
+        expect(spec_helper).to include 'SimpleCov.start do'
       end
 
       context 'running inside the new gem' do