+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # master
 
+* Add default configuration for simplecov in spec_helper.rb
+
 * Display an error message if bundler failed to run or is not present in the system
 
-    * Juan Manuel Ramallo*
+    *Juan Manuel Ramallo*
 
 # 0.1.3